Moving Card to Utils
