+++ goald-frontend/src/components/Group/Widgets/Info/index.jsx
@@ -1,4 +1,4 @@
-import Card from "../Card";
+import Card from "@components/Utils/Card";
 import "./Info.scss"
 
 export default function Info(props) {
@@ -9,7 +9,6 @@ export default function Info(props) {
   } = props
   return (
     <Card
-      title={''}
       children={
         <div className="group-info">
           <div className="group-info_banner"></div>
