+++ goald-frontend/src/components/Utils/Card/Card.scss
